Updating the resultet metadata when remote JDBC metadata changes
As we generate frames, check to see if the metadata changes
for each progressive row. If the metadata changes, then we
end the frame and send along the new metadata for the next
frame.

On the client-side, we manage the transition with a callback
from the Frame FetchIterator when it finds a new signature.
The signature is only propagated until right before the next
batch of rows (e.g. the frame to which the new metadata applies).

Its a little bit ugly how we reach into the resultset to manage
the cursor update, but unfortunately seemingly necessary.